+++ packages/rollup-plugin-html/src/RollupPluginHTMLOptions.ts
@@ -20,6 +20,8 @@ export interface TagAndAttribute {
 export interface RollupPluginHTMLOptions {
   /** HTML file(s) to use as input. If not set, uses rollup input option. */
   input?: string | InputHTMLOptions | (string | InputHTMLOptions)[];
+  /** HTML file glob pattern or patterns to ignore */
+  exclude?: string | string[];
   /** Whether to minify the output HTML. */
   minify?: boolean;
   /** Whether to preserve or flatten the directory structure of the HTML file. */