+++ packages/rollup-plugin-html/src/output/hashInlineScripts.ts
@@ -1,4 +1,4 @@
-import { Node, Document, DefaultTreeElement } from 'parse5';
+import { Document, Element, ParentNode } from 'parse5';
 import {
   findElement,
   findElements,
@@ -13,7 +13,7 @@ import {
 } from '@web/parse5-utils';
 import crypto from 'crypto';
 
-function isMetaCSPTag(node: Node) {
+function isMetaCSPTag(node: Element) {
   if (
     getTagName(node) === 'meta' &&
     getAttribute(node, 'http-equiv') === 'Content-Security-Policy'
@@ -23,7 +23,7 @@ function isMetaCSPTag(node: Node) {
   return false;
 }
 
-function isInlineScript(node: Node) {
+function isInlineScript(node: Element) {
   if (getTagName(node) === 'script' && !hasAttribute(node, 'src')) {
     return true;
   }
@@ -105,7 +105,7 @@ function serializeMetaCSPContent(data: { [key: string]: string[] }): string {
   }, '');
 }
 
-function injectCSPScriptRules(metaCSPEl: DefaultTreeElement, hashes: string[]) {
+function injectCSPScriptRules(metaCSPEl: Element, hashes: string[]) {
   const content = getAttribute(metaCSPEl, 'content');
   if (content) {
     const data = parseMetaCSPContent(content);
@@ -128,7 +128,7 @@ function injectCSPMetaTag(document: Document, hashes: string[]) {
   });
   const head = findNode(document, node => node.nodeName === 'head');
   if (head) {
-    prepend(head, metaTag);
+    prepend(head as ParentNode, metaTag);
   }
 }