+++ packages/test-runner-commands/test/snapshot/snapshotPlugin.test.ts
@@ -1,22 +1,30 @@
 import path from 'path';
 import { runTests } from '@web/test-runner-core/test-helpers';
-import { chromeLauncher } from '@web/test-runner-chrome';
+import { playwrightLauncher } from '@web/test-runner-playwright';
 
 import { snapshotPlugin } from '../../src/snapshotPlugin';
 
-describe.only('snapshotPlugin', function test() {
+describe('snapshotPlugin', function test() {
   this.timeout(20000);
 
   it('passes snapshot tests', async () => {
     await runTests({
       files: [path.join(__dirname, 'browser-test.js')],
-      browsers: [chromeLauncher()],
+      browsers: [
+        playwrightLauncher({ product: 'firefox' }),
+        playwrightLauncher({ product: 'chromium' }),
+        playwrightLauncher({ product: 'webkit' }),
+      ],
       plugins: [snapshotPlugin()],
     });
 
     await runTests({
       files: [path.join(__dirname, 'src', 'nested-test.js')],
-      browsers: [chromeLauncher()],
+      browsers: [
+        playwrightLauncher({ product: 'firefox' }),
+        playwrightLauncher({ product: 'chromium' }),
+        playwrightLauncher({ product: 'webkit' }),
+      ],
       plugins: [snapshotPlugin()],
     });
   });