+++ packages/test-runner/demo/test/__snapshots__/pass-snapshot.test.snap.js
@@ -1,15 +1,16 @@
-export const snapshotsVersion = 1;
+/* @web/test-runner snapshot v1 */
+
 export const snapshots = {}
 
 snapshots["snapshot-a"] = 
 `some snapshot A`;
-// end snapshot snapshot-a
+/* end snapshot snapshot-a */
 
 snapshots["snapshot-b"] = 
 `some snapshot B`;
-// end snapshot snapshot-b
+/* end snapshot snapshot-b */
 
 snapshots["snapshot-c"] = 
 `some snapshot B`;
-// end snapshot snapshot-c
+/* end snapshot snapshot-c */