+++ packages/rollup-plugin-html/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @web/rollup-plugin-html
 
+## 1.9.1
+
+### Patch Changes
+
+- ca749b0e: Update dependency @types/parse5 to v6
+- Updated dependencies [ca749b0e]
+  - @web/parse5-utils@1.3.0
+
+## 1.9.0
+
+### Minor Changes
+
+- 075afcf9: Add `exclude` option. Used e.g. to ignore html file assets when bundling.
+
 ## 1.8.0
 
 ### Minor Changes