+++ packages/test-runner/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @web/test-runner
 
+## 0.13.15
+
+### Patch Changes
+
+- 73681b6d: Allow user config to be partial
+- Updated dependencies [73681b6d]
+  - @web/test-runner-core@0.10.19
+
+## 0.13.14
+
+### Patch Changes
+
+- cb693c71: Use block comments in snapshots to make them work in all browsers
+- Updated dependencies [cb693c71]
+  - @web/test-runner-commands@0.5.6
+
 ## 0.13.13
 
 ### Patch Changes